+++ Epoch/include/Epoch.h
@@ -12,12 +12,15 @@
 
 #include "Epoch/OrthographicCameraController.h"
 
+#include "Epoch/Graphics/Layers/ImGuiLayer.h"
+
 // --------------------- Renderer -----------------------------------------------------------------
 #include "Epoch/Graphics/Renderer.h"
 #include "Epoch/Graphics/RendererCommand.h"
 
 #include "Epoch/Graphics/Buffer.h"
 #include "Epoch/Graphics/Shader.h"
+#include "Epoch/Graphics/Texture.h"
 #include "Epoch/Graphics/VertexArray.h"
 
 #include "Epoch/Graphics/OrthographicCamera.h"