+++ Epoch/include/Epoch/Graphics/RendererAPI.h
@@ -16,7 +16,7 @@ namespace Epoch {
 		virtual void setClearColor(const glm::vec4& color) = 0;
 		virtual void clear() = 0;
 
-		virtual void drawIndexed(const std::shared_ptr<VertexArray>& veretxArray) = 0;
+		virtual void drawIndexed(const Ref<VertexArray>& veretxArray) = 0;
 
 		inline static API getAPI() { return cm_API; }