+++ Epoch/src/Epoch/Graphics/Buffer.cpp
@@ -76,12 +76,12 @@ namespace Epoch {
 	// ----------------- Vertex Buffer ------------------------------------------------------------
 	// --------------------------------------------------------------------------------------------
 
-	VertexBuffer* VertexBuffer::create(float* vertices, uint32_t size)
+	Ref<VertexBuffer> VertexBuffer::create(float* vertices, uint32_t size)
 	{
 		switch (Renderer::getAPI())
 		{
 		case RendererAPI::API::None:			EGE_CORE_ASSERT(false, "None is not yet implemented"); return nullptr;
-		case RendererAPI::API::OpenGL:			return new OpenGLVertexBuffer(vertices, size);
+		case RendererAPI::API::OpenGL:			return std::make_shared<OpenGLVertexBuffer>(vertices, size);
 		default:								EGE_CORE_ASSERT(false, "Unknown render API"); return nullptr;
 		}
 	}
@@ -90,12 +90,12 @@ namespace Epoch {
 	// ----------------- Index Buffer -------------------------------------------------------------
 	// --------------------------------------------------------------------------------------------
 
-	IndexBuffer* IndexBuffer::create(uint32_t* indices, uint32_t count)
+	Ref<IndexBuffer> IndexBuffer::create(uint32_t* indices, uint32_t count)
 	{
 		switch (Renderer::getAPI())
 		{
 		case RendererAPI::API::None:			EGE_CORE_ASSERT(false, "None is not yet implemented"); return nullptr;
-		case RendererAPI::API::OpenGL:			return new OpenGLIndexBuffer(indices, count);
+		case RendererAPI::API::OpenGL:			return std::make_shared<OpenGLIndexBuffer>(indices, count);
 		default:								EGE_CORE_ASSERT(false, "Unknown render API"); return nullptr;
 		}
 	}