+++ Epoch/src/Epoch/Platform/OpenGL/OpenGLRendererAPI.cpp
@@ -7,6 +7,8 @@ namespace Epoch {
 
 	void OpenGLRendererAPI::init()
 	{
+		glEnable(GL_BLEND);
+		glBlendFunc(GL_SRC_ALPHA, GL_ONE_MINUS_SRC_ALPHA);
 	}
 
 	void OpenGLRendererAPI::setClearColor(const glm::vec4& color)