+++ Sandbox/src/Sanbox/Sandbox.cpp
@@ -13,27 +13,28 @@ SandboxLayer::SandboxLayer()
 	: Layer("SandboxLayer"), m_CameraController(16.0f / 9.0f)
 {
 	// Blue Square VAO
-	m_BlueSquareVAO.reset(Epoch::VertexArray::create());
+	m_FlatSquareVAO.reset(Epoch::VertexArray::create());
 
-	float square_vertices[3 * 7] = {
-		-0.5f, -0.5f, 0.0f,
-		 0.5f, -0.5f, 0.0f,
-		 0.5f,  0.5f, 0.0f,
-		-0.5f,  0.5f, 0.0f
+	float square_vertices[5 * 4] = {
+		-0.5f, -0.5f, 0.0f, 0.0f, 0.0f,
+		 0.5f, -0.5f, 0.0f, 1.0f, 0.0f,
+		 0.5f,  0.5f, 0.0f, 1.0f, 1.0f,
+		-0.5f,  0.5f, 0.0f, 0.0f, 1.0f
 	};
 
-	std::shared_ptr<Epoch::VertexBuffer> blueSquareVertexBuffer;
-	blueSquareVertexBuffer.reset(Epoch::VertexBuffer::create(square_vertices, sizeof(square_vertices)));
+	Epoch::Ref<Epoch::VertexBuffer> squareVBO;
+	squareVBO = Epoch::VertexBuffer::create(square_vertices, sizeof(square_vertices));
 
-	blueSquareVertexBuffer->setBufferLayout({
+	squareVBO->setBufferLayout({
 		{ Epoch::ShaderDataType::Float3, "a_Position" },
+		{ Epoch::ShaderDataType::Float2, "a_TextCoord" },
 	});
-	m_BlueSquareVAO->addVertexBuffer(blueSquareVertexBuffer);
+	m_FlatSquareVAO->addVertexBuffer(squareVBO);
 
 	uint32_t square_indices[6] = { 0, 1, 2, 2, 3, 0 };
-	std::shared_ptr<Epoch::IndexBuffer> blueSquareIndexBuffer;
-	blueSquareIndexBuffer.reset(Epoch::IndexBuffer::create(square_indices, sizeof(square_indices) / sizeof(uint32_t)));
-	m_BlueSquareVAO->setIndexBuffer(blueSquareIndexBuffer);
+	std::shared_ptr<Epoch::IndexBuffer> squareIBO;
+	squareIBO = Epoch::IndexBuffer::create(square_indices, sizeof(square_indices) / sizeof(uint32_t));
+	m_FlatSquareVAO->setIndexBuffer(squareIBO);
 
 	// Blue Square Shader
 	std::string flatShaderVertexSrc = R"(
@@ -60,13 +61,22 @@ SandboxLayer::SandboxLayer()
 
 			in vec3 v_Position;
 
+			uniform vec3 u_Color;
+
 			void main()
 			{
-				color = vec4(0.2, 0.3, 0.8, 1.0);
+				color = vec4(u_Color, 1.0);
 			}
 		)";
 
-	m_FlatColorShader.reset(Epoch::Shader::create(flatShaderVertexSrc, flatShaderFragmentSrc));
+	m_FlatColorShader = Epoch::Shader::create(flatShaderVertexSrc, flatShaderFragmentSrc);
+
+	m_TextureShader = Epoch::Shader::create("assets/shaders/Texture.glsl");
+
+	m_CheckerboardTexture = Epoch::Texture2D::create("assets/textures/Checkerboard.png");
+
+	m_FlatColorShader->bind();
+	std::dynamic_pointer_cast<Epoch::OpenGLShader>(m_FlatColorShader)->uploadUniformInt("u_Texture", 0);
 }
 
 void SandboxLayer::onUpdate(Epoch::Timestep ts)
@@ -80,17 +90,21 @@ void SandboxLayer::onUpdate(Epoch::Timestep ts)
 
 	glm::mat4 scale = glm::scale(glm::mat4(1.0f), glm::vec3(0.1f));
 
+	m_FlatColorShader->bind();
+	std::dynamic_pointer_cast<Epoch::OpenGLShader>(m_FlatColorShader)->uploadUniformFloat3("u_Color", m_SquareColor);
+
 	for (int y = 0; y < 20; y++)
 	{
 		for (int x = 0; x < 20; x++)
 		{
 			glm::vec3 pos(x * 0.11f, y * 0.11f, 0.0f);
 			glm::mat4 transform = glm::translate(glm::mat4(1.0f), pos) * scale;
-			Epoch::Renderer::submit(m_FlatColorShader, m_BlueSquareVAO, transform);
+			Epoch::Renderer::submit(m_FlatColorShader, m_FlatSquareVAO, transform);
 		}
 	}
 
-	Epoch::Renderer::submit(m_FlatColorShader, m_BlueSquareVAO);
+	m_CheckerboardTexture->bind();
+	Epoch::Renderer::submit(m_TextureShader, m_FlatSquareVAO, glm::scale(glm::mat4(1.0f), glm::vec3(1.5f)));
 
 	Epoch::Renderer::endScene();
 }
@@ -102,8 +116,8 @@ void SandboxLayer::onEvent(Epoch::Event& e)
 
 void SandboxLayer::onImGuiRender()
 {
-	ImGui::Begin("Text");
-	ImGui::Text("Hello, World");
+	ImGui::Begin("Color");
+	ImGui::ColorEdit3("Square Color", value_ptr(m_SquareColor));
 	ImGui::End();
 }